Consolidate op serialization logic
